requirements.txt
# Core PyTorch and Lightning
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
lightning>=2.0.0

# Medical AI
monai>=1.3.0
monai[all]>=1.3.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Experiment Tracking (Optional)
wandb>=0.15.0

# Data Processing and Visualization
matplotlib>=3.7.0
scikit-learn>=1.3.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.22.0
pandas>=2.0.0

# Additional utilities for Vast.ai
tqdm>=4.65.0
tensorboard>=2.13.0
jupyter>=1.0.0
ipywidgets>=8.0.0

# Performance optimization
numba>=0.57.0
